crosh: Fix issues found by cargo clippy
BUG=None
TEST=cargo test

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4169647
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>

+++ crosh/src/main.rs
@@ -127,7 +127,7 @@ fn handle_cmd(dispatcher: &Dispatcher, args: Vec<String>) -> Result<(), ()> {
         dispatch_cmd(dispatcher, args);
     }
 
-    COMMAND_RUNNING_PID.store(pid as i32, Ordering::Release);
+    COMMAND_RUNNING_PID.store(pid, Ordering::Release);
 
     let mut status: c_int = 1;
     // Safe because status is owned.
